+++ crates/native/build.rs
@@ -194,7 +194,7 @@ fn link_libs() {
     // since 13.0.
     //
     // Should be in `libwebrtc-sys` but has no effect there - weak link
-    // is not being carry overed to `libmedea_flutter_webrtc_native.dylib`
+    // is not being carried over to `libmedea_flutter_webrtc_native.dylib`
     println!("cargo:rustc-link-arg=-Wl,-weak_framework,ScreenCaptureKit");
 
     println!("cargo:rustc-link-lib=framework=AVFoundation");