+++ example/integration_test/webrtc_test.dart
@@ -1362,8 +1362,7 @@ void main() {
 
     var tracks = await getDisplayMedia(caps);
 
-    expect(tracks.length, 2);
-    expect(tracks.where((track) => track.kind() == MediaKind.video).length, 1);
+    expect(tracks.length, 1);
     expect(tracks.where((track) => track.kind() == MediaKind.audio).length, 1);
 
     var audioTrack = tracks.firstWhere(